requirements.txt
comtypes==1.1.10
cycler==0.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.0.25
mysqlclient==2.0.3
numpy==1.21.0
Pillow==8.3.1
protobuf==3.17.3
pyparsing==2.4.7
pypiwin32==223
python-dateutil==2.8.1
pyttsx3==2.90
pywin32==301
six==1.16.0
